Thu, Jan 6, 2022, 5:18 PM +00:00

+++ readability/readability.c
@@ -17,10 +17,10 @@ int main(void)
 
 
     float per_hundred = 100 / count_words(text);
-    int L = (count_letters(text) / count_words(text)) * per_hundred;    //average number of letters per 100 words
-    printf("%i\n", L);
-    int S = (count_sentences(text) / count_words(text)) * per_hundred;  //average number of sentences per 100 words
-    printf("%i\n", S);
+    float L = (count_letters(text) / count_words(text)) * per_hundred;    //average number of letters per 100 words
+    printf("%f\n", L);
+    float S = (count_sentences(text) / count_words(text)) * per_hundred;  //average number of sentences per 100 words
+    printf("%f\n", S);
     int index = 0.0588 * L - 0.296 * S - 15.8;
 
     printf("Grade %i\n", index);
